// src/usecases/movie/interfaces/movieDto.ts
export interface IMovieDTO {
  id: number;
  adult: boolean;
  backdrop_path: string;
  genre_ids: any[];
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}